[CHORE] 확장성 고려해 Manager 로직 개선 (#71)
