+++ Plu-iOS/Plu-iOS/Scenes/OthersAnswer/Manager/OthersAnswerManagerImpl.swift
@@ -9,20 +9,7 @@ import Foundation
 
 
 final class OthersAnswerManagerImpl: OthersAnswerManager {
-    func getAnswers(type: AnswerFilterButtonType) -> OthersAnswer {
-        switch type {
-        case .latest:
-            return OthersAnswer.dummmy()
-        case .mostEmpathy:
-            let dummy = OthersAnswer.dummmy()
-            let answers = dummy.answers.sorted { lhs, rhs in
-                return lhs.empathyCount > rhs.empathyCount
-            }
-            return OthersAnswer(elementType: dummy.elementType,
-                                date: dummy.date,
-                                question: dummy.question,
-                                answersCount: dummy.answersCount,
-                                answers: answers)
-        }
+    func getAnswers(filter: AnswerFilter) async throws -> OthersAnswer {
+        return filter.getOtherAnswers()
     }
 }
