+++ Plu-iOS/Plu-iOS/Scenes/OthersAnswer/Model/AnswerFilterButtonType.swift
@@ -2,7 +2,7 @@
 //  AnswerFilterButtonType.swift
 //  Plu-iOS
 //
-//  Created by 김민재 on 12/6/23.
+//  Created by 김민재 on 12/22/23.
 //
 
 import Foundation
@@ -11,6 +11,15 @@ import Foundation
 enum AnswerFilterButtonType {
     case latest
     case mostEmpathy
+    
+    var object: AnswerFilter {
+        switch self {
+        case .latest:
+            return LatestFilter()
+        case .mostEmpathy:
+            return EmpathyFilter()
+        }
+    }
     
     var title: String {
         switch self {
